refactor: 카드컨테이너의 관심사 분리 및 좋아요 네트워크 요청 권한 부여
